# Conflicts:
#	modules/client-handler/src/main/java/org/apache/ignite/client/handler/requests/table/ClientTableCommon.java
#	modules/client-handler/src/main/java/org/apache/ignite/client/handler/requests/tx/ClientTransactionBeginRequest.java

+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/tree/io/BplusLeafIo.java
@@ -52,7 +52,7 @@ public abstract class BplusLeafIo<L> extends BplusIo<L> {
 
         assertPageType(dstPageAddr);
 
-        copyMemory(srcPageAddr, offset(srcIdx), dstPageAddr, offset(dstIdx), cnt * getItemSize());
+        copyMemory(srcPageAddr, offset(srcIdx), dstPageAddr, offset(dstIdx), cnt * (long) getItemSize());
     }
 
     @Override
